requirements.txt
alembic==1.10.3
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.27.0
backports-cached-property==1.0.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
configargparse==1.5.3
cryptography==40.0.2
deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.87.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
limits==3.4.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pylogbeat==2.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-logstash-async==2.5.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
setuptools==67.6.1
six==1.16.0
sniffio==1.3.0
sqladmin==0.7.0
SQLAlchemy==1.4.47
sqlalchemy-utils==0.41.0
starlette==0.21.0
strawberry-graphql==0.171.1
toml==0.10.2
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchgod==0.8.2
wrapt==1.15.0
wtforms==3.0.1
yarl==1.8.2
zipp==3.15.0
pre-commit==3.3.2
coverage==7.2.2
email-validator==1.3.0
selenium==4.11.2
